[QCF-102] UserBadge 컴포넌트화
