+++ desk/src/components/ui/userBadge/types.ts
@@ -1,0 +1,3 @@
+export type UserBadgeProps = {
+  job: string | undefined
+}
